chore: update dependencies, add shell completion and improve documentation for livetrace CLI
